server/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
bidict==0.21.2
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
crypto==1.4.1
cryptography==3.4.7
Django==3.2.4
django-channels==0.7.0
django-request-logging==0.7.3
djangorestframework==3.12.4
dnspython==1.16.0
enum-compat==0.0.3
eventlet==0.31.1
greenlet==0.4.16
idna==2.10
monotonic==1.5
mypy-extensions==0.4.3
Naked==0.1.31
oauthlib==3.1.1
pathspec==0.8.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.18.0
python-engineio==4.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.5
requests==2.25.1
requests-oauthlib==1.3.0
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
